Fix Plus plant RenderType

+++ src/main/java/com/infinityraider/agricraft/api/v1/plant/IAgriGrowable.java
@@ -38,5 +38,5 @@ public interface IAgriGrowable {
      * @param stage the growth stage
      * @return height of the plant in 1/16ths of a block
      */
-    int getPlantHeight(IAgriGrowthStage stage);
+    double getPlantHeight(IAgriGrowthStage stage);
 }
